update
remove SECP256K1.swift
move remaining SECP256K1 methods to secp256k1 extension for further translation to secp256k1 implementation
